app_client/requirements.txt
paho-mqtt==1.3.1
pep8==1.7.1
pygame==1.9.4.dev0
PyQt5==5.10.1
sip==4.19.8
six==1.11.0
